+++ mycroft/skills/intent_service.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 #
 import time
-from os.path import isfile
 from adapt.context import ContextManagerFrame
 from adapt.engine import IntentDeterminationEngine
 from adapt.intent import IntentBuilder
@@ -26,8 +25,6 @@ from mycroft.util.parse import normalize
 from mycroft.metrics import report_timing, Stopwatch
 from mycroft.skills.padatious_service import PadatiousService
 from .intent_service_interface import open_intent_envelope
-from mycroft.messagebus.client import MessageBusClient
-from mycroft.util import create_daemon
 
 
 class AdaptIntent(IntentBuilder):
@@ -119,7 +116,7 @@ class ContextManager:
                               relevant_frames[i].entities]
             for entity in frame_entities:
                 entity['confidence'] = entity.get('confidence', 1.0) \
-                    / (2.0 + depth)
+                                       / (2.0 + depth)
             context += frame_entities
 
             # Update depth
@@ -184,6 +181,7 @@ class IntentService:
 
         def add_active_skill_handler(message):
             self.add_active_skill(message.data['skill_id'])
+
         self.bus.on('active_skill_request', add_active_skill_handler)
         self.active_skills = []  # [skill_id , timestamp]
         self.converse_timeout = 5  # minutes to prune active_skills
@@ -197,9 +195,11 @@ class IntentService:
         self.bus.on('intent.service.adapt.get', self.handle_get_adapt)
         self.bus.on('intent.service.intent.get', self.handle_get_intent)
         self.bus.on('intent.service.skills.get', self.handle_get_skills)
-        self.bus.on('intent.service.active_skills.get', self.handle_get_active_skills)
+        self.bus.on('intent.service.active_skills.get',
+                    self.handle_get_active_skills)
         self.bus.on('intent.service.adapt.manifest.get', self.handle_manifest)
-        self.bus.on('intent.service.adapt.vocab.manifest.get', self.handle_vocab_manifest)
+        self.bus.on('intent.service.adapt.vocab.manifest.get',
+                    self.handle_vocab_manifest)
 
     def update_skill_name_dict(self, message):
         """
@@ -355,8 +355,8 @@ class IntentService:
                     for utt in combined:
                         _intent = PadatiousService.instance.calc_intent(utt)
                         if _intent:
-                            best = padatious_intent.conf if padatious_intent\
-                                        else 0.0
+                            best = padatious_intent.conf if padatious_intent \
+                                else 0.0
                             if best < _intent.conf:
                                 padatious_intent = _intent
                     LOG.debug("Padatious intent: {}".format(padatious_intent))
@@ -370,7 +370,7 @@ class IntentService:
                               {'intent_type': 'converse'})
                 return
             elif (intent and intent.get('confidence', 0.0) > 0.0 and
-                    not (padatious_intent and padatious_intent.conf >= 0.95)):
+                  not (padatious_intent and padatious_intent.conf >= 0.95)):
                 # Send the message to the Adapt intent's handler unless
                 # Padatious is REALLY sure it was directed at it instead.
                 self.update_context(intent)
@@ -544,7 +544,8 @@ class IntentService:
         # Adapt intent's handler is used unless
         # Padatious is REALLY sure it was directed at it instead.
         padatious_intent = PadatiousService.instance.calc_intent(utterance)
-        if intent is None or (padatious_intent and padatious_intent.conf >= 0.95):
+        if intent is None or (
+                padatious_intent and padatious_intent.conf >= 0.95):
             intent = padatious_intent.__dict__
         self.bus.emit(message.reply("intent.service.intent.reply",
                                     {"intent": intent}))
@@ -555,7 +556,8 @@ class IntentService:
 
     def handle_get_active_skills(self, message):
         self.bus.emit(message.reply("intent.service.skills.reply",
-                                    {"skills": [s[0] for s in self.active_skills]}))
+                                    {"skills": [s[0] for s in
+                                                self.active_skills]}))
 
     def handle_manifest(self, message):
         self.bus.emit(message.reply("intent.service.adapt.manifest",
@@ -564,219 +566,3 @@ class IntentService:
     def handle_vocab_manifest(self, message):
         self.bus.emit(message.reply("intent.service.adapt.vocab.manifest",
                                     {"vocab": self.registered_vocab}))
-
-
-class IntentApi:
-    """
-    Query Intent Service at runtime
-    """
-
-    def __init__(self, bus=None, timeout=5):
-        if bus is None:
-            bus = MessageBusClient()
-            create_daemon(bus.run_forever)
-        self.bus = bus
-        self.timeout = timeout
-        self.bus.on('intent.service.padatious.reply', self._receive_data)
-        self.bus.on('intent.service.adapt.reply', self._receive_data)
-        self.bus.on('intent.service.intent.reply', self._receive_data)
-        self.bus.on('intent.service.skills.reply', self._receive_data)
-        self.bus.on('intent.service.padatious.manifest', self._receive_data)
-        self.bus.on('intent.service.adapt.manifest', self._receive_data)
-        self.bus.on('intent.service.adapt.vocab.manifest', self._receive_data)
-        self.bus.on('intent.service.padatious.entities.manifest', self._receive_data)
-        self._response = None
-        self.waiting = False
-
-    def _receive_data(self, message):
-        self.waiting = False
-        self._response = message.data
-
-    def get_adapt_intent(self, utterance):
-        """ get best adapt intent for utterance """
-        start = time.time()
-        self._response = None
-        self.waiting = True
-        self.bus.emit(Message("intent.service.adapt.get",
-                              {"utterance": utterance},
-                              context={"destination": "intent_service",
-                                       "source": "intent_api"}))
-        while self.waiting and time.time() - start <= self.timeout:
-            time.sleep(0.3)
-        if time.time() - start > self.timeout:
-            LOG.error("Intent Service timed out!")
-            return None
-        return self._response["intent"]
-
-    def get_padatious_intent(self, utterance):
-        """ get best padatious intent for utterance """
-        start = time.time()
-        self._response = None
-        self.waiting = True
-        self.bus.emit(Message("intent.service.padatious.get",
-                              {"utterance": utterance},
-                              context={"destination": "intent_service",
-                                       "source": "intent_api"}))
-        while self.waiting and time.time() - start <= self.timeout:
-            time.sleep(0.3)
-        if time.time() - start > self.timeout:
-            LOG.error("Intent Service timed out!")
-            return None
-        return self._response["intent"]
-
-    def get_intent(self, utterance):
-        """ get best intent for utterance """
-        start = time.time()
-        self._response = None
-        self.waiting = True
-        self.bus.emit(Message("intent.service.intent.get",
-                              {"utterance": utterance},
-                              context={"destination": "intent_service",
-                                       "source": "intent_api"}))
-        while self.waiting and time.time() - start <= self.timeout:
-            time.sleep(0.3)
-        if time.time() - start > self.timeout:
-            LOG.error("Intent Service timed out!")
-            return None
-        return self._response["intent"]
-
-    def get_skills(self):
-        start = time.time()
-        self._response = None
-        self.waiting = True
-        self.bus.emit(Message("intent.service.skills.get",
-                              context={"destination": "intent_service",
-                                       "source": "intent_api"}))
-        while self.waiting and time.time() - start <= self.timeout:
-            time.sleep(0.3)
-        if time.time() - start > self.timeout:
-            LOG.error("Intent Service timed out!")
-            return None
-        return self._response["skills"]
-
-    def get_active_skills(self):
-        start = time.time()
-        self._response = None
-        self.waiting = True
-        self.bus.emit(Message("intent.service.active_skills.get",
-                              context={"destination": "intent_service",
-                                       "source": "intent_api"}))
-        while self.waiting and time.time() - start <= self.timeout:
-            time.sleep(0.3)
-        if time.time() - start > self.timeout:
-            LOG.error("Intent Service timed out!")
-            return None
-        return self._response["skills"]
-
-    def get_adapt_manifest(self):
-        start = time.time()
-        self._response = None
-        self.waiting = True
-        self.bus.emit(Message("intent.service.adapt.manifest.get",
-                              context={"destination": "intent_service",
-                                       "source": "intent_api"}))
-        while self.waiting and time.time() - start <= self.timeout:
-            time.sleep(0.3)
-        if time.time() - start > self.timeout:
-            LOG.error("Intent Service timed out!")
-            return None
-        return self._response["intents"]
-
-    def get_padatious_manifest(self):
-        start = time.time()
-        self._response = None
-        self.waiting = True
-        self.bus.emit(Message("intent.service.padatious.manifest.get",
-                              context={"destination": "intent_service",
-                                       "source": "intent_api"}))
-        while self.waiting and time.time() - start <= self.timeout:
-            time.sleep(0.3)
-        if time.time() - start > self.timeout:
-            LOG.error("Intent Service timed out!")
-            return None
-        return self._response["intents"]
-
-    def get_intent_manifest(self):
-        padatious = self.get_padatious_manifest()
-        adapt = self.get_adapt_manifest()
-        return {"adapt": adapt,
-                "padatious": padatious}
-
-    def get_vocab_manifest(self):
-        start = time.time()
-        self._response = None
-        self.waiting = True
-        self.bus.emit(Message("intent.service.adapt.vocab.manifest.get",
-                              context={"destination": "intent_service",
-                                       "source": "intent_api"}))
-        while self.waiting and time.time() - start <= self.timeout:
-            time.sleep(0.3)
-        if time.time() - start > self.timeout:
-            LOG.error("Intent Service timed out!")
-            return None
-        vocab = {}
-        for voc in self._response["vocab"]:
-            if voc.get("regex"):
-                continue
-            if voc["end"] not in vocab:
-                vocab[voc["end"]] = {"samples": []}
-            vocab[voc["end"]]["samples"].append(voc["start"])
-        return [{"name": voc, "samples": vocab[voc]["samples"]}
-                for voc in vocab]
-
-    def get_regex_manifest(self):
-        start = time.time()
-        self._response = None
-        self.waiting = True
-        self.bus.emit(Message("intent.service.adapt.vocab.manifest.get",
-                              context={"destination": "intent_service",
-                                       "source": "intent_api"}))
-        while self.waiting and time.time() - start <= self.timeout:
-            time.sleep(0.3)
-        if time.time() - start > self.timeout:
-            LOG.error("Intent Service timed out!")
-            return None
-
-        vocab = {}
-        for voc in self._response["vocab"]:
-            if not voc.get("regex"):
-                continue
-            name = voc["regex"].split("(?P<")[-1].split(">")[0]
-            if name not in vocab:
-                vocab[name] = {"samples": []}
-            vocab[name]["samples"].append(voc["regex"])
-        return [{"name": voc, "regexes": vocab[voc]["samples"]}
-                for voc in vocab]
-
-    def get_entities_manifest(self):
-        start = time.time()
-        self._response = None
-        self.waiting = True
-        self.bus.emit(Message("intent.service.padatious.entities.manifest.get",
-                              context={"destination": "intent_service",
-                                       "source": "intent_api"}))
-        while self.waiting and time.time() - start <= self.timeout:
-            time.sleep(0.3)
-        if time.time() - start > self.timeout:
-            LOG.error("Intent Service timed out!")
-            return None
-        entities = []
-        # read files
-        for ent in self._response["entities"]:
-            if isfile(ent["file_name"]):
-                with open(ent["file_name"]) as f:
-                    lines = f.read().replace("(", "").replace(")", "").split("\n")
-                samples = []
-                for l in lines:
-                    samples += [a.strip() for a in l.split("|") if a.strip()]
-                entities.append({"name": ent["name"], "samples": samples})
-        return entities
-
-    def get_keywords_manifest(self):
-        padatious = self.get_entities_manifest()
-        adapt = self.get_vocab_manifest()
-        regex = self.get_regex_manifest()
-        return {"adapt": adapt,
-                "padatious": padatious,
-                "regex": regex}
-
