+++ mycroft/skills/padatious_service.py
@@ -28,7 +28,7 @@ from mycroft.util.log import LOG
 class PadatiousService(FallbackSkill):
     instance = None
 
-    fallback_tight_match = 5   # Fallback priority for the conf > 0.8 match
+    fallback_tight_match = 5  # Fallback priority for the conf > 0.8 match
     fallback_loose_match = 89  # Fallback priority for the conf > 0.5 match
 
     def __init__(self, bus, service):
@@ -60,8 +60,10 @@ class PadatiousService(FallbackSkill):
         self.bus.on('detach_skill', self.handle_detach_skill)
         self.bus.on('mycroft.skills.initialized', self.train)
         self.bus.on('intent.service.padatious.get', self.handle_get_padatious)
-        self.bus.on('intent.service.padatious.manifest.get', self.handle_manifest)
-        self.bus.on('intent.service.padatious.entities.manifest.get', self.handle_entity_manifest)
+        self.bus.on('intent.service.padatious.manifest.get',
+                    self.handle_manifest)
+        self.bus.on('intent.service.padatious.entities.manifest.get',
+                    self.handle_entity_manifest)
 
         # Call Padatious an an early fallback, looking for a high match intent
         self.register_fallback(self.handle_fallback,
@@ -193,11 +195,12 @@ class PadatiousService(FallbackSkill):
                                     {"intents": self.registered_intents}))
 
     def handle_entity_manifest(self, message):
-        self.bus.emit(message.reply("intent.service.padatious.entities.manifest",
-                                    {"entities": self.registered_entities}))
+        self.bus.emit(
+            message.reply("intent.service.padatious.entities.manifest",
+                          {"entities": self.registered_entities}))
 
     # NOTE: This cache will keep a reference to this calss (PadatiousService),
     # but we can live with that since it is used as a singleton.
-    @lru_cache(maxsize=2)   # 2 catches both raw and normalized utts in cache
+    @lru_cache(maxsize=2)  # 2 catches both raw and normalized utts in cache
     def calc_intent(self, utt):
         return self.container.calc_intent(utt)
